Move logging support to backend
