+++ src/backend/logging.ts
@@ -1,5 +1,5 @@
 import { Logger } from 'tslog';
-import { isRunningTests } from './util.js';
+import { DEV, isRunningTests } from '../util.js';
 
 // 0: silly, 1: trace, 2: debug, 3: info, 4: warn, 5: error, 6: fatal
 const trace = 2;
@@ -11,7 +11,7 @@ function getLoggingLevel(): number {
     return error;
   }
 
-  if ('DEV' in process.env && process.env.DEV === 'true') {
+  if (DEV) {
     return trace;
   }
