+++ src/backend/rebench/results.ts
@@ -5,7 +5,7 @@ import { BenchmarkData } from '../../shared/api.js';
 import { Database } from '../db/db.js';
 import { createValidator } from './api-validator.js';
 import { DEBUG } from '../../util.js';
-import { log } from '../../logging.js';
+import { log } from '../logging.js';
 import { completeRequest, startRequest } from '../../perf-tracker.js';
 
 const validateFn: ValidateFunction = DEBUG ? createValidator() : <any>undefined;
