+++ src/timeline-calc.ts
@@ -3,7 +3,7 @@ import { SummaryStatistics } from './stats.js';
 import { robustSrcPath } from './util.js';
 import { Worker } from 'node:worker_threads';
 import { completeRequest, startRequest } from './perf-tracker.js';
-import { log } from './logging.js';
+import { log } from './backend/logging.js';
 
 export interface ComputeRequest {
   runId: number;
