+++ tests/api.test.ts
@@ -4,7 +4,7 @@ import { readFileSync } from 'fs';
 import { ValidateFunction } from 'ajv';
 import { createValidator } from '../src/backend/rebench/api-validator.js';
 import { getDirname } from '../src/util.js';
-import { log } from '../src/logging.js';
+import { log } from '../src/backend/logging.js';
 
 const __dirname = getDirname(import.meta.url);
